server/requirements-dev.txt
-r requirements.txt
black>=24.0.0
ruff>=0.1.0
isort>=5.13.0
mypy>=1.8.0
django-stubs>=4.2.0
djangorestframework-stubs>=3.14.0
